Moved state machine helpers to tests/framework.
